refactor: AtomicInteger 통신 작업 수를 TaskCountManager에서 의존
